pyproject.toml
[project]
name = "henchman"
version = "0.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "pydantic>=2.12.5",
    "python-on-whales>=0.80.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
henchman = "src.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.8.0",
]

[tool.pytest]
addopts = ["-vvv"]
